New list view
Converted the grid view to a list view. Not sure if it's a good idea so
I decided to make a new branch instead. Maybe I will add an option that
allows user to toggle between the two views in the future.
Alternatively, the list view can be reserved for reordering only.