Added simplified list view
This is regarded as simplified as there are some settings missing, such
as reverse mode and deleting. I actually enjoyed the list view, so I
might make it a togglable UI. Also, I am thinking of a solution
regarding the possibility to edit each entry in edit mode. Maybe I will
send a "cancel operation" prop to each counter when the user activates
edit mode, so the user cannot edit the counters in list mode.